requirements.txt
google-adk
opentelemetry-instrumentation-google-genai
pytest